mysite/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.3
Django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
django-filter==25.2
django-modeltranslation==0.19.17
django-nested-admin==4.1.6
django-nested-inline==0.4.6
django-rest-swagger==2.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
openapi-codec==1.3.2
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python-monkey-business==1.1.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
simplejson==3.20.2
sqlparse==0.5.4
uritemplate==4.2.0
urllib3==2.5.0
